requirements.txt
# falcon
falcon==4.0.2
# pillow
Pillow==11.1.0